Move TdiPortManager to its own header file (#33)
- Extract TdiPortManager and PortStatusEvent class definitions from
  the TdiSdeInterface class and move them to their own header file.
  Update references and dependencies as appropriate.
