+++ stratum/hal/lib/tdi/dpdk/dpdk_chassis_manager.cc
@@ -29,7 +29,7 @@
 #include "stratum/hal/lib/tdi/dpdk/dpdk_port_config.h"
 #include "stratum/hal/lib/tdi/dpdk/dpdk_port_constants.h"
 #include "stratum/hal/lib/tdi/dpdk/dpdk_port_manager.h"
-#include "stratum/hal/lib/tdi/tdi_sde_interface.h"
+#include "stratum/hal/lib/tdi/tdi_port_manager.h"
 #include "stratum/lib/macros.h"
 #include "stratum/public/proto/error.pb.h"
 
@@ -37,7 +37,7 @@ namespace stratum {
 namespace hal {
 namespace tdi {
 
-using PortStatusEvent = TdiSdeInterface::PortStatusEvent;
+using PortStatusEvent = TdiPortManager::PortStatusEvent;
 
 ABSL_CONST_INIT absl::Mutex chassis_lock(absl::kConstInit);
