+++ stratum/hal/lib/tdi/es2k/es2k_port_manager_mock.h
@@ -18,7 +18,7 @@ class Es2kPortManagerMock : public Es2kPortManager {
 
   MOCK_METHOD(
       ::util::Status, RegisterPortStatusEventWriter,
-      (std::unique_ptr<ChannelWriter<TdiSdeInterface::PortStatusEvent>> writer),
+      (std::unique_ptr<ChannelWriter<PortStatusEvent>> writer),
       (override));
 
   MOCK_METHOD(::util::Status, UnregisterPortStatusEventWriter, (), (override));
