+++ stratum/hal/lib/tdi/tofino/tofino_port_manager.h
@@ -8,7 +8,7 @@
 #include <memory>
 
 #include "absl/synchronization/mutex.h"
-#include "stratum/hal/lib/tdi/tdi_sde_interface.h"
+#include "stratum/hal/lib/tdi/tdi_port_manager.h"
 
 // Suppress clang errors
 #undef LOCKS_EXCLUDED
@@ -18,14 +18,14 @@ namespace stratum {
 namespace hal {
 namespace tdi {
 
-class TofinoPortManager : public TdiSdeInterface::TdiPortManager {
+class TofinoPortManager : public TdiPortManager {
  public:
   TofinoPortManager();
   virtual ~TofinoPortManager() {}
 
   // TdiPortManager public methods
   ::util::Status RegisterPortStatusEventWriter(
-      std::unique_ptr<ChannelWriter<TdiSdeInterface::PortStatusEvent>> writer)
+      std::unique_ptr<ChannelWriter<PortStatusEvent>> writer)
       LOCKS_EXCLUDED(port_status_event_writer_lock_);
   ::util::Status UnregisterPortStatusEventWriter()
       LOCKS_EXCLUDED(port_status_event_writer_lock_);
@@ -98,8 +98,8 @@ class TofinoPortManager : public TdiSdeInterface::TdiPortManager {
 
   // Writer to forward the port status change message to. It is registered
   // by chassis manager to receive SDE port status change events.
-  std::unique_ptr<ChannelWriter<TdiSdeInterface::PortStatusEvent>>
-    port_status_event_writer_ GUARDED_BY(port_status_event_writer_lock_);
+  std::unique_ptr<ChannelWriter<PortStatusEvent>> port_status_event_writer_
+    GUARDED_BY(port_status_event_writer_lock_);
 };
 
 }  // namespace tdi
